zitaku

+++ RealtimeDataAggregation/src/apps/count/agent/aggregate/handler/UpdateAggregateAgentHandler.java
@@ -32,7 +32,7 @@ public class UpdateAggregateAgentHandler  extends MessageHandler{
         TxID tx = getTx();
         long updateData = 0;
         for(UserData user : (List<UserData>)updateMsg.msgdata){
-            updateData =  updateData + (Integer)user.data;
+            updateData =  updateData + user.data;
         }
         
         agent.setData(tx, agent.getData(tx)+updateData);
